requirements.txt
fastapi
uvicorn
gunicorn
pydantic
pydantic-settings
qdrant-client
redis
boto3
openai
python-dotenv
langchain
langchain-openai
httpx
streamlit
requests
tqdm
pymupdf
pandas
python-multipart